requirements.txt
flask
gunicorn
pandas
numpy
sklearn
pyYAML
-e .